+++ app/src/main/AndroidManifest.xml
@@ -9,7 +9,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".MainActivity"
+            android:name=".Controller.MainActivity"
             android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -18,20 +18,20 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".GameActivity"
-            android:parentActivityName=".MainActivity"
+            android:name=".Controller.GameActivity"
+            android:parentActivityName=".Controller.MainActivity"
             android:screenOrientation="portrait">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value=".MainActivity" />
+                android:value=".Controller.MainActivity" />
         </activity>
         <activity
-            android:name=".EndActivity"
-            android:parentActivityName=".EndActivity"
+            android:name=".Controller.EndActivity"
+            android:parentActivityName=".Controller.EndActivity"
             android:screenOrientation="portrait">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value=".GameActivity" />
+                android:value=".Controller.GameActivity" />
         </activity>
     </application>
