+++ app/src/main/java/javalab/pk/saper/Controller/EndActivity.java
@@ -1,12 +1,13 @@
-package javalab.pk.saper;
+package javalab.pk.saper.Controller;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
+
+import javalab.pk.saper.R;
 
 public class EndActivity extends AppCompatActivity {
