+++ app/src/main/java/javalab/pk/saper/Controller/MainActivity.java
@@ -1,10 +1,12 @@
-package javalab.pk.saper;
+package javalab.pk.saper.Controller;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
+
+import javalab.pk.saper.R;
 
 public class MainActivity extends AppCompatActivity {
