--- app/src/main/java/javalab/pk/saper/Pole3.java
@@ -1,12 +1,0 @@
-package javalab.pk.saper;
-
-/**
- * Created by Kamil on 2017-04-09.
- */
-
-public class Pole3 extends PoleNumer {
-    @Override
-    public int action() {
-        return R.mipmap.number_3;
-    }
-}
