--- app/src/main/java/javalab/pk/saper/Pole7.java
@@ -1,12 +1,0 @@
-package javalab.pk.saper;
-
-/**
- * Created by khaeemm on 26.04.2017.
- */
-
-public class Pole7 extends PoleNumer {
-    @Override
-    public int action() {
-        return R.mipmap.number_7;
-    }
-}
