--- app/src/main/java/javalab/pk/saper/PoleNumer.java
@@ -1,8 +1,0 @@
-package javalab.pk.saper;
-
-/**
- * Created by khaeemm on 27.04.2017.
- */
-
-public abstract class PoleNumer extends Pole {
-}
