+++ app/src/main/res/layout/activity_main.xml
@@ -9,7 +9,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="javalab.pk.saper.MainActivity">
+    tools:context="javalab.pk.saper.Controller.MainActivity">
 
 
     <LinearLayout
